+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 3.09.2018
 
+* [Android] Sunshine: Send Data 
 * [Android] Sunshine: Add a new Activity
 
 #### 2.09.2018